requirements.txt
bcrypt==4.1.3
blinker==1.8.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
jwt==1.3.1
MarkupSafe==2.1.5
pycparser==2.22
PyMySQL==1.1.0
python-decouple==3.8
pytz==2024.1
Werkzeug==3.0.2
